+++ js/create-full-photo.js
@@ -1,11 +1,12 @@
 import { isEscKeydown } from './util.js';
 
+const UPLOADED_COMMENTS_AMOUNT = 5;
+
 const pageBody = document.body;
 const fullPhoto = document.querySelector('.big-picture');
 
 const photoAddress = fullPhoto.querySelector('.big-picture img');
 const photoLikes = fullPhoto.querySelector('.likes-count');
-const photoCommentsCount = document.querySelector('.comments-count');
 const photoDescription = document.querySelector('.social__caption');
 
 const closeFullPhotoButton = fullPhoto.querySelector('.big-picture__cancel');
@@ -13,7 +14,9 @@ const closeFullPhotoButton = fullPhoto.querySelector('.big-picture__cancel');
 const photoComments = fullPhoto.querySelector('.social__comments');
 const photoComment = photoComments.querySelector('.social__comment');
 
-const commentsLoader = fullPhoto.querySelector('.comments-loader');
+const socialCommentCount = fullPhoto.querySelector('.social__comment-count');
+const photoCommentsCount = document.querySelector('.comments-count');
+const commentsLoaderButton = fullPhoto.querySelector('.comments-loader');
 
 
 const createComments = ({avatar, name, message}) => {
@@ -35,6 +38,27 @@ const addComments = (comments) => {
   photoComments.append(commentFragment);
 };
 
+//
+let startCommentsCount = 0;
+let commentsData = [];
+let commentsCount = 0;
+
+const renderMoreComments = () => {
+  const moreComments = commentsData.slice(startCommentsCount, startCommentsCount + UPLOADED_COMMENTS_AMOUNT);
+  addComments(moreComments);
+  startCommentsCount = startCommentsCount + UPLOADED_COMMENTS_AMOUNT;
+
+  if (commentsData.length <= startCommentsCount) {
+    commentsLoaderButton.classList.add('hidden');
+  }
+  commentsCount = startCommentsCount;
+  if (startCommentsCount > commentsData.length) {
+    commentsCount = commentsData.length;
+  }
+  socialCommentCount.innerHTML = `${commentsCount} из ${commentsData.length} комментариев`;
+};
+//
+
 const createFullPhoto = ({url, likes, description, comments}) => {
   photoAddress.src = url;
   photoLikes.textContent = likes;
@@ -43,16 +67,21 @@ const createFullPhoto = ({url, likes, description, comments}) => {
 
   document.addEventListener('keydown', onFullPhotoEscKeydown);
 
-  addComments(comments);
-
   fullPhoto.classList.remove('hidden');
-  commentsLoader.classList.add('hidden');
   pageBody.classList.add('modal-open');
+
+  commentsLoaderButton.classList.remove('hidden');
+  commentsData = comments;
+  startCommentsCount = 0;
+  renderMoreComments();
 };
+
+commentsLoaderButton.addEventListener('click', () => {
+  renderMoreComments();
+});
 
 const closeFullPhoto = () => {
   fullPhoto.classList.add('hidden');
-  commentsLoader.classList.remove('hidden');
   pageBody.classList.remove('modal-open');
 };
