+++ js/main.js
@@ -1,7 +1,7 @@
 import { publishedPhotos } from './data.js';
 import { renderMiniatures } from './rendering-miniatures.js';
-import './upload-photo-form.js';
 import { openPhotoUploadForm } from './upload-photo-form.js';
+import './form-validate.js';
 
 renderMiniatures(publishedPhotos);
 openPhotoUploadForm();
