+++ js/upload-photo-form.js
@@ -1,8 +1,4 @@
-import { checkoutTextLength, isEscKeydown } from './util.js';
-
-const PHOTO_DESCRIPTION_MAX_LENGTH = 140;
-const HASHTAG_REGULAR_EXPRESSION = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
-const MAX_TAGS = 5;
+import { isEscKeydown } from './util.js';
 
 const pageBody = document.body;
 
@@ -58,71 +54,5 @@ const openPhotoUploadForm = () => {
   uploadFormFileInput.addEventListener('change', openUploadForm);
 };
 
-//Валидация
-
-//Изначально поставил ограничения по длине в разметке в index, но тогда не выскакивает предупреждение о максимально допустимой длине
-const validatePhotoDescriptionLength = (string) => checkoutTextLength(string, PHOTO_DESCRIPTION_MAX_LENGTH);
-
-const validateHashtagsNames = (data) => {
-  if (data.length > 0) {
-    const hashtagsArray = data.split(' ');
-    return(hashtagsArray.every((hashtag) => HASHTAG_REGULAR_EXPRESSION.test(hashtag)));
-  }
-  return true;
-};
-
-const validateHashtagsAmount = (data) => {
-  const hashtagsArray = data.split(' ');
-  return(hashtagsArray.length <= MAX_TAGS);
-};
-
-const validateHashtagsDublicates = (data) => {
-  const hashtagsArray = data.split(' ');
-  const uniqueHashtagsArray = [...new Set(hashtagsArray.map((hashtag) => hashtag.toLowerCase()))];
-  return (hashtagsArray.length === uniqueHashtagsArray.length);
-};
-
-const pristine = new Pristine(uploadForm, {
-  classTo: 'img-upload__field-wrapper',
-  errorClass: 'img-upload__item--invalid',
-  successClass: 'img-upload__item--valid',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextTag: 'div'
-});
-
-//Проверка максимальной длины строки описания фотографии
-pristine.addValidator(
-  photoDescriptionField,
-  validatePhotoDescriptionLength,
-  'Длина комментария не может составлять больше 140 символов'
-);
-
-//Проверка именования Тега на "маску" разрешенного именования
-pristine.addValidator(
-  hashtagsField,
-  validateHashtagsNames,
-  'Хэш-тег должен начинается с символа #, состоять из букв и чисел, не содержать пробелы, спецсимволы, содержать от 2 до 20 символов'
-);
-
-//Проверка на максимольное допустимое количество Тегов
-pristine.addValidator(
-  hashtagsField,
-  validateHashtagsAmount,
-  `Максимальное количство хэш-тегов не должно превышать ${MAX_TAGS}`
-);
-
-//Проверка Тегов на отсутсвие их повторений
-pristine.addValidator(
-  hashtagsField,
-  validateHashtagsDublicates,
-  'Хэш-теги не должны повторяться'
-);
-
-uploadForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  if (pristine.validate()) {
-    uploadForm.submit();
-  }
-});
 
 export { openPhotoUploadForm };
